+++ app/project/bootstrappers/http/views/Builders.php
@@ -7,26 +7,26 @@ use Project\HTTP\Views\Builders\Edit;
 use Project\HTTP\Views\Builders\Home;
 use Project\HTTP\Views\Builders\Master;
 use Opulence\Applications\Bootstrappers\Bootstrapper;
-use Opulence\Views\Factories\ITemplateFactory;
+use Opulence\Views\Factories\IViewFactory;
 
 class Builders extends Bootstrapper
 {
     /**
      * Registers view builders to the factory
      *
-     * @param ITemplateFactory $templateFactory The template factory to use
+     * @param IViewFactory $templateFactory The template factory to use
      */
-    public function run(ITemplateFactory $templateFactory)
+    public function run(IViewFactory $templateFactory)
     {
-        $templateFactory->registerBuilder("Master.php", function()
+        $templateFactory->registerBuilder("Master.fortune", function()
         {
             return new Master();
         });
-        $templateFactory->registerBuilder("Home.php", function()
+        $templateFactory->registerBuilder("Home.fortune", function()
         {
             return new Home();
         });
-        $templateFactory->registerBuilder("Edit.php", function()
+        $templateFactory->registerBuilder("Edit.fortune", function()
         {
             return new Edit();
         });