+++ app/project/bootstrappers/http/views/View.php
@@ -4,12 +4,12 @@
  */
 namespace Project\Bootstrappers\HTTP\Views;
 use Opulence\Files\FileSystem;
-use Opulence\Framework\Bootstrappers\HTTP\Views\Template as BaseBootstrapper;
+use Opulence\Framework\Bootstrappers\HTTP\Views\View as BaseBootstrapper;
 use Opulence\IoC\IContainer;
 use Opulence\Views\Caching\Cache;
 use Opulence\Views\Caching\ICache;
 
-class Template extends BaseBootstrapper
+class View extends BaseBootstrapper
 {
     /**
      * Gets the view cache