+++ app/project/http/controllers/Page.php
@@ -6,23 +6,23 @@ namespace Project\HTTP\Controllers;
 use Opulence\HTTP\Responses\Response;
 use Opulence\Routing\Controller;
 use Opulence\Views\Compilers\ICompiler;
-use Opulence\Views\Factories\ITemplateFactory;
+use Opulence\Views\Factories\IViewFactory;
 
 class Page extends Controller
 {
-    /** @var ICompiler The template compiler to use */
+    /** @var ICompiler The view compiler to use */
     protected $compiler = null;
-    /** @var ITemplateFactory The factory to use to create templates */
-    protected $templateFactory = null;
+    /** @var IViewFactory The factory to use to create views */
+    protected $viewFactory = null;
 
     /**
-     * @param ICompiler $compiler The template compiler to use
-     * @param ITemplateFactory $templateFactory The factory to use to create templates
+     * @param ICompiler $compiler The view compiler to use
+     * @param IViewFactory $viewFactory The factory to use to create views
      */
-    public function __construct(ICompiler $compiler, ITemplateFactory $templateFactory)
+    public function __construct(ICompiler $compiler, IViewFactory $viewFactory)
     {
         $this->compiler = $compiler;
-        $this->templateFactory = $templateFactory;
+        $this->viewFactory = $viewFactory;
     }
 
     /**
@@ -32,9 +32,9 @@ class Page extends Controller
      */
     public function showEditPage()
     {
-        $this->template = $this->templateFactory->create("Edit.php");
+        $this->view = $this->viewFactory->create("Edit.fortune");
 
-        return new Response($this->compiler->compile($this->template));
+        return new Response($this->compiler->compile($this->view));
     }
 
     /**
@@ -42,11 +42,11 @@ class Page extends Controller
      */
     public function showHTTPError($statusCode)
     {
-        $this->template = $this->templateFactory->create("HTTPError.php");
-        $this->template->setVar("title", $statusCode . " Error");
-        $this->template->setTag("errorTitle", $statusCode . " Error");
+        $this->view = $this->viewFactory->create("HTTPError.fortune");
+        $this->view->setVar("title", $statusCode . " Error");
+        $this->view->setVar("errorTitle", $statusCode . " Error");
 
-        return new Response($this->compiler->compile($this->template), $statusCode);
+        return new Response($this->compiler->compile($this->view), $statusCode);
     }
 
     /**
@@ -56,8 +56,8 @@ class Page extends Controller
      */
     public function showHomePage()
     {
-        $this->template = $this->templateFactory->create("Home.php");
+        $this->view = $this->viewFactory->create("Home.fortune");
 
-        return new Response($this->compiler->compile($this->template));
+        return new Response($this->compiler->compile($this->view));
     }
 }