+++ configs/console/bootstrappers.php
@@ -4,7 +4,7 @@
  */
 use Project\Bootstrappers\Console\Commands\Commands as ProjectCommands;
 use Project\Bootstrappers\HTTP\Routing\Router;
-use Project\Bootstrappers\HTTP\Views\Template;
+use Project\Bootstrappers\HTTP\Views\View;
 use Opulence\Framework\Bootstrappers\Console\Commands\Commands as OpulenceCommands;
 use Opulence\Framework\Bootstrappers\Console\Requests\Requests;
 use Opulence\Framework\Bootstrappers\Console\Composer\Composer;
@@ -14,6 +14,6 @@ return [
     OpulenceCommands::class,
     Requests::class,
     Composer::class,
-    Template::class,
+    View::class,
     ProjectCommands::class
 ];