+++ configs/http/bootstrappers.php
@@ -5,15 +5,15 @@
 use Project\Bootstrappers\HTTP\Routing\Router;
 use Project\Bootstrappers\HTTP\Sessions\Session;
 use Project\Bootstrappers\HTTP\Views\Builders;
-use Project\Bootstrappers\HTTP\Views\Template;
+use Project\Bootstrappers\HTTP\Views\View;
 use Opulence\Framework\Bootstrappers\HTTP\Requests\Request;
-use Opulence\Framework\Bootstrappers\HTTP\Views\TemplateFunctions;
+use Opulence\Framework\Bootstrappers\HTTP\Views\ViewFunctions;
 
 return [
     Request::class,
     Router::class,
-    TemplateFunctions::class,
+    View::class,
     Session::class,
-    Template::class,
+    ViewFunctions::class,
     Builders::class
 ];